refactor: move key-store module under standalone
