+++ src/main/standalone/key-store.ts
